+++ src/main/java/com/nighthawk/spring_portfolio/mvc/binary/BinaryGenerate.java
@@ -5,7 +5,7 @@ import java.util.Random;
 
 public class BinaryGenerate {
     @Id
-    public static double[] getBinary() {
+    public static int[] getBinary() {
         Random rand = new Random();
         double decimal = rand.nextDouble()*255;
         decimal = Math.floor(decimal);
@@ -13,9 +13,9 @@ public class BinaryGenerate {
         while (binaryString.length() < 8) {
             binaryString = "0" + binaryString;
         }
-        double[] bools = new double[2];
-        bools[0] = Long.parseLong(binaryString);
-        bools[1] = decimal;
+        int[] bools = new int[2];
+        bools[0] = Integer.parseInt(binaryString);
+        bools[1] = (int) decimal;
         return bools;
     };
     private int number;
